Rename wizards to tools
