+++ templates/default/root.html.twig
@@ -7,7 +7,7 @@
         <p class="lead">
             {% if communityVersions|length > 0 %}<a href="#download" class="btn btn-lg btn-primary">Download TYPO3</a>{% endif %}
             <a href="#try-online" class="btn btn-lg btn-light">Try TYPO3</a>
-            <a href="{{ path('wizards_sitepackage') }}" class="btn btn-lg btn-light">Create a Sitepackage</a>
+            <a href="{{ path('tools_sitepackage') }}" class="btn btn-lg btn-light">Create a Sitepackage</a>
         </p>
     {% endframe %}
