+++ templates/layout.html.twig
@@ -14,7 +14,7 @@
         gtag(\'config\', \'G-JTSKLNF4S0\');
     ') }}
 {% endblock %}
-{% block headline %}Try the new <a href="{{ path('wizards_sitepackage') }}" style="color:#515151">Sitepackage Builder</a> now and create a Sitepackage in 2 easy steps.{% endblock %}
+{% block headline %}Try the new <a href="{{ path('tools_sitepackage') }}" style="color:#515151">Sitepackage Builder</a> now and create a Sitepackage in 2 easy steps.{% endblock %}
 {% block footer %}
     {% frame with { color: 'primary' } %}
         <div class="card-group-element card-group-element-align-left card-group-element-columns-2">
