+++ templates/tools/layout.html.twig
@@ -4,7 +4,7 @@
 
 {% block footer %}
 
-    {% include 'wizards/partials/privacy-footer.html.twig' %}
+    {% include 'tools/partials/privacy-footer.html.twig' %}
     {{ parent() }}
 
 {% endblock %}
