+++ templates/tools/sitepackage/configure.html.twig
@@ -1,4 +1,4 @@
-{% extends 'wizards/layout.html.twig' %}
+{% extends 'tools/layout.html.twig' %}
 
 {% block title %}Create your own TYPO3 Sitepackage{% endblock %}
